+++ lib/glimmer-dsl-opal/samples/hello/hello_table.rb
@@ -206,13 +206,13 @@ class HelloTable
         table_column {
           text 'Game Date'
           width 150
-#           sort_property :date # ensure sorting by real date value (not `game_date` string specified in items below)
+          sort_property :date # ensure sorting by real date value (not `game_date` string specified in items below)
 #           editor :date_drop_down, property: :date_time
         }
         table_column {
           text 'Game Time'
           width 150
-#           sort_property :time # ensure sorting by real time value (not `game_time` string specified in items below)
+          sort_property :time # ensure sorting by real time value (not `game_time` string specified in items below)
 #           editor :time, property: :date_time
         }
         table_column {